==> 01 get ids.ipynb <==
# **Obtención de los ids disponibles en [The Materials Project](https://next-gen.materialsproject.org/)**

## *Objetivo:* Iterar sobre los ids desde cero hasta dos millones, accediendo a la base de datos de Materials Project para realizar un request y verificar si el id iterado existe. Si el id se encuentra en la base de datos, se registra en un archivo .csv para su posterior tratamiento; de lo contrario, se omite.

### 1. Se cargan librerías y se declaran variables

### Se carga libreria m-api ###
from mp_api.client import MPRester

### Se declaran variables ###
key="Inserta tu APIKEY aqui"
mpr = MPRester(key)

### 2. Se realiza request del id iterado

### Se realiza iteracion ###
with open('datos/ids.csv','w') as file:
  for i in range(20):
    docs = mpr.summary.search(material_ids=['mp-'+str(i)])
    if len(docs)!=0:
      file.write('mp-'+str(i)+'\n')
    i+=1
print('FINISHED')
