==> 03 filtro datos.ipynb <==
# **Filtro de datos**

## *Objetivo:* Aplica filtros utilizando máscaras en el DataFrame que contiene los datos crudos. Este proceso permite reducir el conjunto de datos a aquellos materiales que cumplen con las siguientes condiciones: su valor de IM es igual a cero, el coeficiente de absorción óptica (CA) no es nulo y el valor de band gap (BG) se encuentra entre 0.95 eV y 3 eV.

### 1. Se cargan librerías y se declaran variables

### Se carga libreria y se define###
import pandas as pd
file='datos/datos_crudos.csv'

### 2. Se realiza el filtrado aplicando máscaras en el DataFrame y se guardan los resultados

### Se realizan filtros individuales para testeo###
df=pd.read_csv(file)
df=df.set_index('id')
print('Metales:',df[df['IM']==True].shape[0])
print('Metales con CA:',df[df['IM']==True].dropna().shape[0])
print('No Metales:',df[df['IM']==False].shape[0])
print('No Metales con CA:',df[df['IM']==False].dropna().shape[0])
print('No Metales con CA y 0.95eV< BG <3eV:',df[(df['IM']==False)&((df['BG']>=0.95)&(df['BG']<=3))].dropna().shape[0])
print('Total de materiales:',
      df[df['IM']==True].shape[0]+
      df[df['IM']==False].shape[0])

### Se realiza filtrado en conjunto y se guardan los datos ###
df_clean=df[(df['IM']==False)&((df['BG']>=0.95)&(df['BG']<=3))].dropna()
df_clean.head()
#df_clean.to_csv('datos/datos_clean.csv', index=False) #Descomentar línea si es necesario
