==> 04 preparacion datos.ipynb <==
# **Preparación de datos**

## *Objetivo:* Realiza cambio de variable *str* a *int* para los valores Structure; elimina las columnas innecesarias del DataFrame y se estandarizan los datos de BG, Structure y CA.

### 1. Se cargan librerías y se declaran variables

### Se importan librerías ###
import pandas as pd
from sklearn import preprocessing
file='datos/datos_clean.csv'

### 2. Se realiza cambio de variable *str* a *int* para la variable Structure

 ### Se realiza el cambio de varaible en Structure ###
df = pd.read_csv(file)
cambios={
    'Trigonal':5,
    'Cubic':1,
    'Tetragonal':2,
    'Hexagonal':4,
    'Orthorhombic':3,
    'Monoclinic':6
}
df_replace=df.replace(cambios)

### 3. Se eliminan columnas innecesarias del DataFrame

### Se elimna columnas que no son de interes ###
df_replace=df_replace.drop(columns=['id','No Ele','IM','Is BG Dir','a','b','c','alfa','beta','gamma'])
df_replace.to_csv('datos/df_replace.csv', index=False)

### Se seleccionan caracteristicas para llevar a cabo el agrupamiento ###
features = ['BG','Structure','CA']

### 4. Se realiza la estandarización de datos

### Se estandarizan los features ###
df_standardized = preprocessing.scale(df_replace[features])
X = df_standardized
#np.save('datos/X.npy', X) #Descomentar línea si es necesario
