==> 05 optimizacion de hiperparametros.ipynb <==
# **Optimización de parámetros para K-means**

## *Objetivo:* Optimiza los hiperparámetros para mediante GridSearchCV, evaluando diferentes combinaciones de parámetros. La declaración de param_grid incluye un rango de valores para cada hiperparámetro, y el proceso prueba todas las combinaciones posibles usando validación cruzada (5 particiones) para seleccionar la mejor configuración. 

### 1. Se cargan librerías y se declaran variables

### Se cargan librerias ###
from sklearn.cluster import KMeans
from sklearn.model_selection import GridSearchCV
import numpy as np
import json

### Se cargan los datos e
X = np.load('datos/X.npy')

### 2. Se declaran los valores de los parámetros a iterar

param_grid = {
    'n_clusters': range(2, 31),
    'init': ['k-means++', 'random'],
    'n_init': [10,30,50],
    'max_iter': [100, 200, 300],
    'tol': [0.0001, 0.001, 0.01],
    'algorithm': ['lloyd', 'elkan'],
    'random_state': [0, 42, 84, 126]
}

### 3. Se llama al algoritmo K-means para iterar los parámetros y obtener el mejor resultado

### Se crea el objeto K-Means ###
kmeans = KMeans(random_state=42)

### Se declara grid search ###
grid_search = GridSearchCV(kmeans, param_grid=param_grid, cv=5, n_jobs=-1)

### Se ajusta grid search con los datos entandarizados###
grid_search.fit(X)

### Se imprimen resultados ###
#with open('datos/grid_search.json', 'w') as file:
    #json.dump(grid_search.best_params_, file) #Descomentar lineas si es necesario
print("Best hyperparameters: ", grid_search.best_params_)

Best hyperparameters:  {'algorithm': 'lloyd', 'init': 'random', 'max_iter': 100, 'n_clusters': 29, 'n_init': 10, 'random_state': 42, 'tol': 0.001}
